+++ src/main/java/Albaid/backend/domain/contract/repository/ContractRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.List;
 
 public interface ContractRepository extends JpaRepository<Contract, Integer> {
-    List<Contract> findByMemberId(Long memberId);
+    List<Contract> findByMemberId(Integer memberId);
 
 
     @Query("SELECT c FROM Contract c WHERE c.member.id = :memberId")
